pyproject.toml
[project]
name = "3gpp-fastapi"
version = "0.1.0"
description = "The API to serve data extracted from the 3GPP state machine extractor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "psycopg[binary,pool]>=3.2.9",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ruff>=0.11.11",
]
